Extra allocation for rooted trees. 1. In PhyloTree::ensureCentralPartialParsimonyIsAllocated,     bumped up the vector_count by 2 (just in case the tree is     rooted) (but I worry, later optimisation code may break if     the tree is rooted, since it doesn't know to treat the branch     too the root as special). 2. Declared PhyloTree members, pars_block_size and     total_parsimony_mem_size as uint64_t. 3. Removed some commented out code in     PhyloSuperTreePlen::initializeAllPartialLh (not a material change).

+++ tree/phylotree.h
@@ -2631,9 +2631,9 @@ protected:
             the main memory storing all partial parsimony states for all neighbors of the tree.
             The variable partial_pars in PhyloNeighbor will be assigned to a region inside this variable.
      */
-    UINT*  central_partial_pars;
-    size_t pars_block_size;          //in UINTs
-    size_t total_parsimony_mem_size; //in UINTs
+    UINT*    central_partial_pars;
+    uint64_t pars_block_size;          //in UINTs
+    uint64_t total_parsimony_mem_size; //in UINTs
 
     virtual void reorientPartialLh(PhyloNeighbor* dad_branch, PhyloNode *dad);
